+++ sbi-spec/CHANGELOG.md
@@ -7,7 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [Unreleased]
 
 ### Added
+
 - pmu: add config flags with bitflags in chapter 11
+- fwft: add support for FWFT extension in chapter 18
 
 ### Modified